requirements.txt
streamlit
rembg
pillow
qrcode